Patch to make /etc/pointercal relative to systemconfdir
While /etc/ts.conf is always relative to systemconfdir (set during ./
configure), the /etc/pointercal file is hard-coded to /etc/pointercal.
For users who wish to prefix their installation this is troublesome,
and impossible if they cannot/do not which to modify the root /etc.
The patch below resolves this by making /etc/pointercal relative to
the systemconfdir set during ./configure (or defaults to prefix/etc as
it should). This mirrors the behavior of tslib's /etc/ts.conf file.

Signed-off-by: Daniel Jabbour <daniel@laptouchinc.com>
Signed-off-by: Chris Larson <clarson@kergoth.com>