Fix current values for input-raw
The input driver only sends new ABS_X and ABS_Y events when the value changes,
even after a pen up and a pen down.

If we have the same y value after releasing and touching again, there will be
no ABS_Y event sent and the sample returns at an EV_SYN the value x,0.

So we don't change the current values at pen_up, but a flag that sets the sample
values to 0.

Signed-off-by: Teresa Gamez <t.gamez@phytec.de>
Signed-off-by: Jan Weitzel <j.weitzel@phytec.de>
Signed-off-by: Chris Larson <chris_larson@mentor.com>