cy8mrln-palmpre: improve interpolation till the edges
At the edges the pressure is reduced because not the whole finger is on the
screen. Normally the pressure is about 50 and if it's not, then the finger is
not completly on the screen and we have to simulate the real peak with a pressure
of (default_pressure - actual_pressure).

Also this patch introduces two new config parameters:
pressure: for the default_pressure (default 60)
gesture_height: for the height of the gesturearea in fields (default 1)

Signed-off-by: Thomas Zimmermann <zimmermann@vdm-design.de>