plugins: variance: immediately accept the first touch sample
The first sample of a touch gesture (after pressure was 0) should be
passed on immediately. Many touch devices only serve one sample when
touching the device without moving. We can't lose that sample.

This wasn't much of a problem with resistive touchscreens that continously
serve input events. Recent devices don't do that anymore.

The variance filter does it's work as soon as 2 samples are seen. It
doesn't make sense earlier anyways.

Fixes Issue #10