README: use markdown file extension
That's a little much, but I'll try this, let's see if we keep it for
the release:

 * have an empty README file so we don't have to set GNU autotools to
   'foreign'
 * have README.markdown as the real readme file.

It *is* our documentation after all and should be as appealing and as
easy to read as possible. I don't want to turn users away from the project
because the description is harder to read than it is with other projects.