include manpages for commands, API and configuration
This starts a man pages project in the doc directory. Some manpages found
in old Debian distributions were used as templates, are greatly modified
and updated. New pages are added too.

In contrast to former man pages for tslib, they aren't generated, but
written by hand and maintained and distributed with the project directly.