fix memory allocation to be portable
or as Coverity puts it:

suspicious_sizeof: Passing argument /* sizeof (struct ts_sample_mt **) */ to function malloc and then casting the return value to struct ts_sample_mt ** is suspicious. In this particular case sizeof (struct ts_sample_mt **) happens to be equal to sizeof (struct ts_sample_mt *), but this is not a portable assumption.