input-raw: add debug warning on SYN_DROPPED
This is the very least we have to do. It would probably make more sense to
write an libevdev based input raw module instead of spending much time on
reinventing the wheel here.

But of course it wouldn't hurt to properly handle SYN_DROPPED directly and
here's some more information on how one would do that:
https://www.freedesktop.org/software/libevdev/doc/latest/syn_dropped.html

Fixes #69