README: add a manual TOC
our chapters shouldn't change much anyways. Going forward this is our documentation
process:

we experiment with new things in the github wiki, and if it fits and looks good,
we'll include it in the README and remove it from the wiki (link to the README).

Remember, the Github wiki will someday be lost. The README won't.