libts: remove package version from library name - introduce libts.so.0
this removes the package version from the libts library name, resulting in
the simple file setup

libts.so -> libts.so.0.2.6
libts.so.0 -> libts.so.0.2.6
libts.so.0.2.6

previously, there was only libts.so staying the same for users because the
package version name was part of the soname.

Now users have a more secure option to use libts.so.0 that ensures backwards
compatibility.

This also updates the library version to indicate an ABI addition (libts.so.0)
and removes the -rc1 from the package version name to be prepared for the
release.