input-raw: ts_read_mt: fix returning actually read samples
We need to return any number of samples we have actually read, smaller than
the number of samples the user *can* read (or requests).

This is especially important when used in non-blocking mode, i.e.

	ts_setup(NULL, 1);

We only have support for ts_read_mt() in input-raw and this fixes the problem
there.

This also adds the number of samples read to the ts_print_mt program.