galax: version 210: use ABS_MT_DISTANCE to find pressure 0
Appearently there are version 210 of the 0eef:0001 devices that won't send
BTN_TOUCH 0 or ABS_MT_TRACKING_ID -1 on pen-up.

Instead they use ABS_MT_DISTANCE 0 during touching and 1 to indicate
"pressure 0" or pen-up.

This implements just that, for this specific device only. This is not at
all the corrent driver-behaviour according to the protocol after all.