input-raw: support all eGalax devices too, including multitouch
This uses code from galax-raw and includes support for most of it's eGalax
devices in input-raw:

both ts_read() and ts_read_mt() should support all versions of eGalax devices
which means

   multitouch capable eGalax devices can now use ts_read_mt() for full
   multitouch support.

Also it sets up the mechanism for working around special broken devices in
the future. Let's hope we won't have to.