add ts_libversion() API for retrieving version information
and update all users that need version information. This should make the
library more usable in the long term. Especially applications can now
simply check for TSLIB_VERSION_MT to be defined to know that they can
use ts_read_mt().

Most importantly, version_num in the new struct that ts_libversion returns
(filled, as it is statically available) is an integer that *always* grows
from version to version. It does *not* include the tarball package version,
but the *library* versions:

<8 bits major number> | <8 bits minor number> | <8 bits patch number>

And as always, the major number should stay 0, so effectively there are 16
significant bits, to differentiate between all future versions of tslib (or
better: libts).