add dmc_dus3000-raw module - support DMC DUS Series over UART
The DUS3000 (but others seem to be supported here too) has a UART and
USB interface. This driver only uses UART and singletouch.

Furthermore the device' calibration capabilities are not used, in order
to keep the driver generic. Simply use the linear module and ts_calibrate
instead.

I add this driver because there is really no support in Linux for these device
right now. Of course, the module can be used on FreeBSD or other platforms
too, which makes the inclusion more ok :)

To enable it, do

	./configure --enable-dmc_dus3000

Many thanks to the donor of this driver module!