ts_version: remove age from the version struct. (Last minute API fix)
The age value would be totally unused right now. And we'd have to add
a mechanism to determinde it anyways. But I really don't see the need
to have it at all. Strictly speaking, the library version number would
we enough for users. The features bits are of value I think, but will
most probably all there is ever needed to use the library correctly.