waveshare: fix usage of free()
There are a few problems in this if() path. First, 0 is a valid file
descriptor and free(-1) is ignored, so the check must go away.

Second, ts_close() frees the tsdev, so no need to free it by hand.

Third, (still open issue!) we should never call open() or close()
directly. ts_open() and ts_close() is the way to go.

Signed-off-by: Martin Kepplinger <martink@posteo.de>