tests: improve evemu_test.sh
This greatly improves evemu_test.sh in that everything is now automated.
We make use of a file naming convention described in README.md, read
filter test files, call evemu-device, start tslib using the new device,
and finally play the test recording to the device. The use watches
the filtered resulting samples.

So what would be left to automate is the evaluation. The criteria for
passing the test however is filter-specific. We could save an additional
third "expected" file, record the actual result, and compare.

Also, running all available tests in a row will be left to a seperate
script for now.

Signed-off-by: Martin Kepplinger <martin.kepplinger@ginzinger.com>