autoconf: disable dependency tracking by default
Since make distclean and out-of-source builds don't work when using
the (enabled by default) dependency tracking, let's disable it by
default. This also fixes make distcheck.

This describes our problem: https://stackoverflow.com/q/21609580/6435680

If this is not a bug in autotools, we'll need a real solution here. But
if we don't refactor much, we don't lose much during development. Plus,
we can simply use --enable-dependency-tracking.

Closes #111

Signed-off-by: Martin Kepplinger <martin.kepplinger@ginzinger.com>