ts_setup: add touchscreen device detection instead of picking event0
This removes trying to open /dev/input/event0 unconditionally in case
1. no device is given on the commandline
2. TSLIB_TSDEVICE is not set
3. /dev/input/ts or /dev/input/touchscreen is not there

Instead we now scan /dev/input/event* until we find a touchscreen device
and take it, in this case.

This should basically deprecate all config settings we have regarding
the touchscreen device: as long as there is only one, it just works.

Signed-off-by: Martin Kepplinger <martin.kepplinger@ginzinger.com>