ts_uinput: add ts_setup_ext() to scan /dev/input/ for a touchscreen
We can't take ts_setup() here because we need to get back the chosen
input device number. We implement an extended version here. Again:

Instead of trying to open /dev/input/event0, we now scan /dev/input/event*
for a touchscreen device. That is, if we aren't forced at a different
one (/dev/input/ts exists, or TSLIB_TSDEVICE is set, for example)

Signed-off-by: Martin Kepplinger <martin.kepplinger@ginzinger.com>