ts_uinput: fix invalid memory access
During the non-blocking test read we don't yet know the actual size of
samples memory we will need. So use a temporary testsample of one slot
here.

We need ts_setup_ext() to find the input device, which in turn is needed
for setup_uinput() which we then need to have the number of slots we
need. While we could go all the way and finding that out only for our
test read, it's not necessary.

Signed-off-by: Martin Kepplinger <martin.kepplinger@ginzinger.com>