ts.conf: remove variance filter from commented-out list
The variance filter plugin is currently deprecated as is has no support
for multitouch; it has no implementation for read_mt(). It seems unlikely
that it will become much better maintained in the future.

We have quite some changes in our example ts.conf file for the current release
already. So: do away with it and don't unnecessarily promote it. We should
have enough filters to build a good user experience without it and it's
more user-friendly to promote only filters that have full multitouch support.

Signed-off-by: Martin Kepplinger <martin.kepplinger@ginzinger.com>