input-raw: mark BTN_TOUCH code valid
In case we have *only* BTN_TOUCH in a sync-frame, we had not marked the
frame as valid and thus missed pen-up for device that *don't* report
pressure or tracking-id or pen-up in any other way. So fix this.

Also, this removes the obsolete handling of BTN_LEFT for ts_read_mt().
It did nothing, and driver's shouldn't use it anyways.

Fixes #120

Signed-off-by: Martin Kepplinger <martin.kepplinger@ginzinger.com>