median: fix pen_down propagation for read_mt()
The median filter silently removed pen-down because that falls under
the "collecting phase" it needs for it's calculation and sets those
invalid.

We now generate a pen-down as soon as the collecting phase is over
and samples are propagated.

This should fix single-touch touchscreen without pressure information
to be used with ts_uinput.

Fixes #133