ts_calibrate / linear / calibfile: add support for rotated calibration
We add the rotation value that was used as ts_calibrate -r option, to the
calibfile format (new value at the end). Other than that, screen-rotation
is technically ignored during calibration.

We read the rotation value in the "linear" module and apply it after the
transformations we already have.

Also, we a "rot" module parameter to the linear module to be able to
overwrite what is read from the calibfile. This is important for
environments that do touch-rotation at a higher level anyways.

Closes #128