input-raw / ts_uinput: fix MT type A device detection and conversion
ts_uinput always output a type B device (if MT). This fixes input-raw's
type A detection and data-generation for type B data.

That said, we do not go as far as libmtdev (http://bitmath.org/code/mtdev/)
goes with A to B conversion. For us, samples that come in first (in a frame)
are slot 0, and so on. No real tracking of touch points using distance
calculations based on previous states. Still, this should result in a totally
usable input stream for most users.

Also, ts_uinput now actually does what it claims to do. If it detects
a MT device, is creates a type B devices, no matter what. For type A
devices, it defaults to 5 slots (usable touch points), if not overwritten
by the user (ts_uinput -s commandline option)

Fixes #120