CMake: use GNUInstallDirs module
This not only provides sensible defaults for many systems (e.g. correct man
location on *BSD, automatic selection of lib64 if needed etc.), it also allows
the user to override all these things from the CMake commandline.

As a side effect this fixes installation of ts.conf and friends, which ended up
in /usr/etc if configured with a prefix of /usr until now.