CMake: fix ENABLE_TOOLS
The option command has 2 mandatory arguments, the option name and a help text.
The third, optional, argument is the default value of the switch, and defaults
to off. This resulted in tools not being built by default. Move the switch one
level up and skip the whole directory if nothing is to do.